requirements.txt
tensorflow-gpu==1.14
numpy
matplotlib
python-telegram-bot
python-dateutil
tqdm